% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_debug.R
\name{geom_debug}
\alias{geom_debug}
\title{Debug ggplot2}
\usage{
geom_debug(
  debug = "head",
  type = NULL,
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  inherit.aes = TRUE
)
}
\arguments{
\item{debug}{a function which conducted with argument \code{coords}. \cr \cr
coords is value that have already been transformed to the plot scales.
Default: \code{head}}

\item{type}{\code{group} or \code{panel}, group which draws geom with
\code{draw_group} draws collective geoms that display multiple observations
with one geometric object and \code{panel} which draws geom with
\code{draw_panel} draws individual geoms that display a distinct graphical
object for each observation (row). Default: \code{group}}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}. These are often
aesthetics, used to set an aesthetic to a fixed value, like colour = "red"
or size = 3. They may also be parameters to the paired geom/stat.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Debug ggplot2 by operate data with a function.
}
