% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_grid.R
\name{geom_draw}
\alias{geom_draw}
\title{Layer with Grid or Function}
\usage{
geom_draw(
  draw = grid::nullGrob(),
  type = NULL,
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  inherit.aes = TRUE
)
}
\arguments{
\item{draw}{Either a \link[grid:grid.grob]{grob} object or a function which accepts
two arguments (\code{data} and \code{coords}) and returns a
\link[grid:grid.grob]{grob}. \cr \cr \code{data} contains value from parameter
\code{data}, coords is value that have already been transformed to the plot
scales. \cr \cr It is important to note that the columns names for the data
and coords come from the aesthetic mappings in the \code{ggplot2} plot.}

\item{type}{\code{group} or \code{panel}, group which draws geom with
\code{draw_group} draws collective geoms that display multiple observations
with one geometric object and \code{panel} which draws geom with
\code{draw_panel} draws individual geoms that display a distinct graphical
object for each observation (row). Default: \code{group}}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}. These are often
aesthetics, used to set an aesthetic to a fixed value, like \code{colour = "red"} or \code{size = 3}. They may also be parameters to the paired geom/stat.
Other arguments of function \code{draw} can be passed here.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
a ggplot2 layer object
}
\description{
Draw ggplot2 layer with a grod or function.
}
\details{
If you want to combine the functionality of multiple geoms it can
usually be achieved by preparing the data for each of the geoms inside the
\verb{draw_*()} call and send it off to the different geoms, collecting the
output in a \code{\link[grid:grid.grob]{grid::gList}} (a list of grobs) if the call is \code{draw_group()}
or a \code{\link[grid:grid.grob]{grid::gTree}} (a grob containing multiple children grobs) if the
call is \code{draw_panel()}.
}
\examples{
gggrid_text <- grid::textGrob(
  "gggrid",
  x = c(0, 0, 0.5, 1, 1),
  y = c(0, 1, 0.5, 0, 1),
  hjust = c(0, 0, 0.5, 1, 1),
  vjust = c(0, 1, 0.5, 0, 1)
)
ggplot2::ggplot(data.frame(x = 1, y = 2))+
  geom_draw(gggrid_text)
}
