% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdraw.R
\docType{package}
\name{ggdraw}
\alias{ggdraw}
\title{ggdraw: Draw with grob or function in ggplot2}
\description{
Draw with grob or function in ggplot2. This package is inspired
by \href{https://github.com/pmur002/gggrid}{gggrid} but adjusted with a lot of
personal convention. See \href{https://www.stat.auckland.ac.nz/~paul/Reports/gggrid/gggrid.html}{gggrid: it's g-g-great!}
}
\seealso{
Useful links:
\itemize{
\item \url{https://ggplot2.tidyverse.org}
\item \url{https://github.com/tidyverse/ggplot2}
\item \url{https://github.com/Yunuuuu/ggdraw}
\item \href{https://www.stat.auckland.ac.nz/~paul/Reports/gggrid/gggrid.html}{gggrid: it's g-g-great!}
}
}
\author{
Yun \email{yunyunpp96@outlook.com}
}
